django/requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
django-cors-headers==3.10.0  
googlemaps==4.5.3
python-dateutil==2.8.2
django-rest-knox==4.1.0
